Move ShutdownHook back to tillerinobot core
